fix: reconnect the client on update failure
Client goes through the full flow of reconnects when `watchCh` becomes
nil. If the `watchCtx` gets canceled, it aborts the `watch()` loop, but
new connection is not established, this includes not sending a `Hello`
request.

This is visible in our server metrics when after a migration some
connections are lingering in the "old" cluster doing just
"AffiliateUpdate" calls.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>